# LEAF-4725 Optimize query used in Inbox: Organize by Roles