## Summary
This improves load performance for the Inbox's "Organize by Roles" mode by ~27% in testing, and up to ~80% in a site containing 2.5k pending requests:
```
                              │ before.txt  │              after.txt              │
                              │   sec/op    │   sec/op     vs base                │
Inbox_adminActionableRoles-24   138.8m ± 2%   101.3m ± 4%  -27.06% (p=0.000 n=10)
```

The improvement is achieved by leveraging the `data.metadata` column to retrieve information (first/last name, email), which avoids needing a query for each unique user in the set.

Since this closes the remaining advantage of the old Inbox, this PR also discourages use of the old Inbox. Visitors are redirected to the current one. In the event something is overlooked, the old inbox remains accessible at a new link: `PORTAL_URL/?a=inbox_old`

## Impact
This makes changes to queries for Actionable records, which are primarily used in the Inbox and Report Builder.

Since the system is unlikely to report "Needs Reassignment" in the Inbox for new records, a different approach should be taken to highlight records assigned to inactive accounts. In the meantime, process owners should use the "Unresolved Requests" feature to ensure timely completion.

## Testing
New Automated Tests: https://github.com/department-of-veterans-affairs/LEAF-Automated-Tests/pull/72